spring5-recipe-app/src/main/java/com/manos/spring5recipeapp/controllers/RecipeController.java
package com.manos.spring5recipeapp.controllers;

import com.manos.spring5recipeapp.services.RecipeService;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;

@Controller
@Slf4j
public class RecipeController {

    @Autowired
    RecipeService recipeService;

    @RequestMapping("recipe/show/{id}")
    public String showById(@PathVariable String id,Model model){
        model.addAttribute("recipe",recipeService.findById(new Long(id)));
        return "recipe/show";
    }
}
